内核：升级至 6.12.103 并重构 Photonicat 移植补丁集
将 Photonicat BSP 从 6.1.106 主线基线升级到 6.12.103，并针对新内核重新构建
overlay/defconfig 补丁集，移除现已冗余、已损坏或已并入上游主线的补丁。

改动内容：
- 在 build-base.sh 中将内核版本由 6.1.106 提升至 6.12.103
- 基于 6.12 基线重写 photonicat_defconfig（大幅重写）
- 退役不再适用或已无必要的上游/Rockchip BSP 补丁，包括 Motorcomm
  yt8521/yt8531 PHY 驱动、stmmac SGMII/QSGMII 支持、xpcs、rk3328/rk3399
  开发板支持、dfi/ddr 时钟 devfreq、rockchip hw_random、CAN-FD、gicv3 ITS
  变通补丁，以及若干 rk3568 HDMI/rfkill/电源关闭补丁
- 删除已由 6.12 内核自带提供的 kernel-overlay 源码（rk3328_dmc、
  rockchip-rng、rk3328 dram dt-bindings/dtsi）
- 更新剩余的自定义补丁（mipi-lcd、rk3399 超频、rfkill-gpio-neo、
  led-mode、fake-battery）以及适配 6.12 的 rk3568-photonicat DTS/pm 驱动
- Debian rootfs：新增 cleanup_rootfs_caches() 并在 minimal/custom/full
  构建流程末尾调用，使 .tar.gz / rootfs-debian-* 产物保持缓存干净，
  同时构建过程仍可正常 apt install
- 新增 gmac0 SGMII 恢复与 6.12 bring-up 问题的规划文档

---

kernel: upgrade to 6.12.103 and rebase Photonicat bring-up patches

Upgrade the Photonicat BSP from the mainline 6.1.106 base to 6.12.103 and
rebuild the overlay/defconfig patch set against the newer kernel, removing
patches that are now redundant, broken, or already merged upstream.

Changes:
- Bump kernel version 6.1.106 -> 6.12.103 in build-base.sh
- Rebase photonicat_defconfig against the 6.12 base (major rewrite)
- Retire upstream/obsolete Rockchip BSP patches that no longer apply or are
  no longer needed, including Motorcomm yt8521/yt8531 PHY drivers,
  stmmac SGMII/QSGMII support, xpcs, rk3328/rk3399 board support,
  dfi/dram-clock devfreq, rockchip hw_random, CAN-FD, gicv3 ITS workaround,
  and assorted rk3568 HDMI/rfkill/Power-Off patches
- Drop kernel-overlay sources now provided by the 6.12 tree (rk3328_dmc,
  rockchip-rng, rk3328 dram dt-bindings/dtsi)
- Update remaining custom patches (mipi-lcd, rk3399 overclock, rfkill-gpio-neo,
  led-mode, fake-battery) and the rk3568-photonicat DTS/pm driver for 6.12
- Debian rootfs: add cleanup_rootfs_caches() and call it at the end of the
  minimal/custom/full builds so .tar.gz / rootfs-debian-* outputs are left
  cache-clean while builds can still apt install
- Add planning docs for gmac0 SGMII restore and 6.12 bring-up issues

+++ patches/kernel-overlay/drivers/power/supply/photonicat-pm.c
@@ -751,7 +751,7 @@ static size_t pcat_pm_uart_receive_parse(struct pcat_pm_data *pm_data,
 	return used_size;
 }
 
-static int pcat_pm_uart_serdev_receive_buf(
+static size_t pcat_pm_uart_serdev_receive_buf(
 	struct serdev_device *serdev, const u8 *buf, size_t count)
 {
 	struct pcat_pm_data *pm_data = serdev_device_get_drvdata(serdev);
